Move pole, zero, evalfr, freqresp to lti module
