+++ control/lti.py
@@ -14,6 +14,9 @@ timebaseEqual()
 
 from numpy import absolute, real
 
+__all__ = ['issiso', 'timebase', 'timebaseEqual', 'isdtime', 'isctime',
+           'pole', 'zero', 'evalfr', 'freqresp']
+
 class LTI:
     """LTI is a parent class to linear time-invariant (LTI) system objects.
 
@@ -189,3 +192,164 @@ def isctime(sys, strict=False):
 
     # Got passed something we don't recognize
     return False
+
+def pole(sys):
+    """
+    Compute system poles.
+
+    Parameters
+    ----------
+    sys: StateSpace or TransferFunction
+        Linear system
+
+    Returns
+    -------
+    poles: ndarray
+        Array that contains the system's poles.
+
+    Raises
+    ------
+    NotImplementedError
+        when called on a TransferFunction object
+
+    See Also
+    --------
+    zero
+
+    Notes
+    -----
+    This function is a wrapper for StateSpace.pole and
+    TransferFunction.pole.
+
+    """
+
+    return sys.pole()
+
+
+def zero(sys):
+    """
+    Compute system zeros.
+
+    Parameters
+    ----------
+    sys: StateSpace or TransferFunction
+        Linear system
+
+    Returns
+    -------
+    zeros: ndarray
+        Array that contains the system's zeros.
+
+    Raises
+    ------
+    NotImplementedError
+        when called on a TransferFunction object or a MIMO StateSpace object
+
+    See Also
+    --------
+    pole
+
+    Notes
+    -----
+    This function is a wrapper for StateSpace.zero and
+    TransferFunction.zero.
+
+    """
+
+    return sys.zero()
+
+def evalfr(sys, x):
+    """
+    Evaluate the transfer function of an LTI system for a single complex
+    number x.
+
+    To evaluate at a frequency, enter x = omega*j, where omega is the
+    frequency in radians
+
+    Parameters
+    ----------
+    sys: StateSpace or TransferFunction
+        Linear system
+    x: scalar
+        Complex number
+
+    Returns
+    -------
+    fresp: ndarray
+
+    See Also
+    --------
+    freqresp
+    bode
+
+    Notes
+    -----
+    This function is a wrapper for StateSpace.evalfr and
+    TransferFunction.evalfr.
+
+    Examples
+    --------
+    >>> sys = ss("1. -2; 3. -4", "5.; 7", "6. 8", "9.")
+    >>> evalfr(sys, 1j)
+    array([[ 44.8-21.4j]])
+    >>> # This is the transfer function matrix evaluated at s = i.
+
+    .. todo:: Add example with MIMO system
+    """
+    if issiso(sys):
+        return sys.horner(x)[0][0]
+    return sys.horner(x)
+
+def freqresp(sys, omega):
+    """
+    Frequency response of an LTI system at multiple angular frequencies.
+
+    Parameters
+    ----------
+    sys: StateSpace or TransferFunction
+        Linear system
+    omega: array_like
+        List of frequencies
+
+    Returns
+    -------
+    mag: ndarray
+    phase: ndarray
+    omega: list, tuple, or ndarray
+
+    See Also
+    --------
+    evalfr
+    bode
+
+    Notes
+    -----
+    This function is a wrapper for StateSpace.freqresp and
+    TransferFunction.freqresp.  The output omega is a sorted version of the
+    input omega.
+
+    Examples
+    --------
+    >>> sys = ss("1. -2; 3. -4", "5.; 7", "6. 8", "9.")
+    >>> mag, phase, omega = freqresp(sys, [0.1, 1., 10.])
+    >>> mag
+    array([[[ 58.8576682 ,  49.64876635,  13.40825927]]])
+    >>> phase
+    array([[[-0.05408304, -0.44563154, -0.66837155]]])
+
+    .. todo::
+        Add example with MIMO system
+
+        #>>> sys = rss(3, 2, 2)
+        #>>> mag, phase, omega = freqresp(sys, [0.1, 1., 10.])
+        #>>> mag[0, 1, :]
+        #array([ 55.43747231,  42.47766549,   1.97225895])
+        #>>> phase[1, 0, :]
+        #array([-0.12611087, -1.14294316,  2.5764547 ])
+        #>>> # This is the magnitude of the frequency response from the 2nd
+        #>>> # input to the 1st output, and the phase (in radians) of the
+        #>>> # frequency response from the 1st input to the 2nd output, for
+        #>>> # s = 0.1i, i, 10i.
+    """
+
+    return sys.freqresp(omega)
